i64 -> Int

+++ src/dijkstra.hpp
@@ -1,24 +1,24 @@
 #include <bits/stdc++.h>
 using namespace std;
-using i64 = long long;
+using Int = long long;
 template <typename T, typename U>
 inline bool chmin(T &a, U b) {
   return a > b and ((a = std::move(b)), true);
 }
 
-const i64 kBig = 1e16;
+const Int kBig = 1e16;
 
 template <class T>
 using MinHeap = priority_queue<T, vector<T>, greater<T>>;
 
 struct Edge {
   int to;
-  i64 cost;
+  Int cost;
 };
 using Graph = std::vector<std::vector<Edge>>;
 
 struct State {
-  i64 cost;
+  Int cost;
   int node;
 };
 bool operator>(const State &x, const State &y) { return x.cost > y.cost; }
@@ -26,9 +26,9 @@ bool operator>(const State &x, const State &y) { return x.cost > y.cost; }
 // Returns min distance from the start node to each node (if exists).
 auto search_shortest_path(const Graph &g, int start, int goal) {
   const int n = g.size();
-  auto mincost = vector(n, kBig);
+  auto mincost = vector(n, (Int)kBig);
   MinHeap<State> que;
-  auto push = [&](i64 cost, int node) -> bool {
+  auto push = [&](Int cost, int node) -> bool {
     if (chmin(mincost[node], cost)) {
       que.push(State{cost, node});
       return true;
@@ -51,7 +51,7 @@ auto search_shortest_path(const Graph &g, int start, int goal) {
 }
 
 struct GridState {
-  i64 cost;
+  Int cost;
   int r;
   int c;
 };
@@ -67,7 +67,7 @@ auto search_shortest_path_on_grid(const vector<string> &g,
   const array<int, 4> dx = {0, 1, 0, -1}, dy = {1, 0, -1, 0};
   auto mincost = vector(H, vector(W, kBig));
   MinHeap<GridState> que;
-  auto push = [&](i64 cost, int r, int c) -> bool {
+  auto push = [&](Int cost, int r, int c) -> bool {
     if (r < 0 or r >= H or c < 0 or c >= W) return false;
     if (g[r][c] == '#') return false;
     if (not chmin(mincost[r][c], cost)) return false;
